:truck: [#4608] Move objects API specific configuration to own file
