+++ src/openforms/js/components/admin/form_design/variables/prefill/types.js
@@ -1,0 +1,6 @@
+import PropTypes from 'prop-types';
+
+export const ErrorsType = PropTypes.oneOfType([
+  PropTypes.arrayOf(PropTypes.oneOfType([PropTypes.string, PropTypes.array])),
+  PropTypes.string,
+]);
